src/requirements.txt
PyYAML==5.3.1
python-telegram-bot==13.11